+++ src/model/output/decoded/greedy.rs
@@ -10,6 +10,7 @@ use super::SpanOutput;
 /// See section 4.2 of <https://aclanthology.org/2022.umios-1.1.pdf>
 pub struct GreedySearch {
     flat_ner: bool,
+    dup_label: bool,
     multi_label: bool,
 }
 
@@ -19,14 +20,14 @@ impl GreedySearch {
     /// Arguments:
     /// * `flat_ner`: if `true`, a span is not allowed not embed another one
     /// * `multi_label`: if `true`, the same span can belong to multiple classes
-    pub fn new(flat_ner: bool, multi_label: bool) -> Self {
-        Self { flat_ner, multi_label }
+    pub fn new(flat_ner: bool, dup_label: bool, multi_label: bool) -> Self {
+        Self { flat_ner, dup_label, multi_label }
     }
 
     /// Perform greedy search
     /// 
     /// Note: spans are supposed to be sorted by start, and then end, offsets.
-    pub fn search(spans: &[Span], flat_ner: bool, multi_label: bool) -> Vec<Span> {
+    pub fn search(&self, spans: &[Span]) -> Vec<Span> {
         if spans.is_empty() {
             return Vec::new();
         }
@@ -38,7 +39,7 @@ impl GreedySearch {
         while next < spans.len() {
             let prev_span = spans.get(prev).unwrap();
             let next_span = spans.get(next).unwrap();
-            if !Self::overlapping(prev_span, next_span, flat_ner, multi_label) {
+            if self.accept(prev_span, next_span) {
                 result.push(prev_span.clone());
                 prev = next;
             }
@@ -54,14 +55,23 @@ impl GreedySearch {
         result
     }
 
-    /// Returns `true` iif:
-    /// * one span overlaps the other one
-    /// * and the two spans don't share the same offsets, unless `multi_label` is set to `true`
-    /// * and neither span is nested in into another, unless `flat_ner` is set to `false`
-    fn overlapping(s1: &Span, s2: &Span, flat_ner: bool, multi_label: bool) -> bool {
-        if s1.same_offsets(s2) { !multi_label }
-        else if flat_ner && (s1.is_nested_in(s2) || s2.is_nested_in(s1)) { false }
-        else { s1.overlaps(s2) }
+    /// Returns `true` iif the span is valid wrt. the provided flags.
+    /// 
+    /// Namely:
+    /// * All overlapping spans are forbiden if `flat_ner` is `true`.
+    /// * Otherwise, `dup_label=true` allows for overlapping spans with the *same* label.
+    /// * And `multi_label=true` allows for overlapping spans with *different* labels.
+    /// 
+    /// The checks are only relative to the previous span, as we expect them to be sorted by offset.    
+    fn accept(&self, s1: &Span, s2: &Span) -> bool {                
+        // if there is no overlap, we accept immediately
+        if s1.is_disjoint(s2) { true }
+        // otherwise, if `flat_ner=true` we reject immediately
+        else if self.flat_ner { false }
+        // otherwise, if the labels are the same, we accept overlaps only if `dup_label=true`
+        else if !self.dup_label && s1.class().eq(s2.class()) { false }
+        // otherwise the labels are different, so we accept overlaps only if `multi_label=true`
+        else { self.multi_label }
     }
 }
 
@@ -71,7 +81,7 @@ impl Composable<SpanOutput, SpanOutput> for GreedySearch {
     fn apply(&self, input: SpanOutput) -> Result<SpanOutput> {        
         let spans = input.spans
             .iter()
-            .map(|s| Self::search(s, self.flat_ner, self.multi_label))
+            .map(|s| self.search(s))
             .collect();
         Ok(SpanOutput::new(input.texts, input.entities, spans))
     }