+++ src/model/params.rs
@@ -11,9 +11,11 @@
 pub struct Parameters {
     /// Probability threshold (default: 0.5)
     pub threshold: f32,    
-    /// Flat NER means that an entity can not embed another one (default: true)
+    /// Setting this parameter to `true` means that no entity can overlap with another one (default: true)
     pub flat_ner: bool,
-    /// Multi-label means that the same span can belong to multiple classes (default: false)
+    /// If `flat_ner=false`, setting this parameter to `true` means that overlapping spans can belong to the *same* class (default: false)
+    pub dup_label: bool,
+    /// If `flat_ner=false`, setting this parameter to `true` means that overlapping spans can belong to *different* classes (default: false)
     pub multi_label: bool,    
     /// For span mode, maximum span width (default: 12)
     pub max_width: usize,
@@ -28,7 +30,8 @@ impl Default for Parameters {
             0.5, 
             12, 
             Some(512),
-            true, 
+            true,             
+            false,
             false,
         )
     }
@@ -36,12 +39,13 @@ impl Default for Parameters {
 
 impl Parameters {
     /// New configuration specifying every parameter
-    pub fn new(threshold: f32, max_width: usize, max_length: Option<usize>, flat_ner: bool, multi_label: bool) -> Self {
+    pub fn new(threshold: f32, max_width: usize, max_length: Option<usize>, flat_ner: bool, dup_label: bool, multi_label: bool) -> Self {
         Self { 
             threshold, 
             max_width, 
             max_length,
             flat_ner,
+            dup_label,
             multi_label,
         }
     }
@@ -63,6 +67,11 @@ impl Parameters {
 
     pub fn with_flat_ner(mut self, flat_ner: bool) -> Self {
         self.flat_ner = flat_ner;
+        self
+    }
+
+    pub fn with_dup_label(mut self, dup_label: bool) -> Self {
+        self.dup_label = dup_label;
         self
     }