+++ src/text/span.rs
@@ -40,17 +40,22 @@ impl Span {
         self.probability
     }
 
-    /// returns `true` if this span is nested inside (or equals) the given span
+    /// returns `true` iif this span is nested inside (or equals) the given span
     pub fn is_nested_in(&self, other: &Span) -> bool {
         self.start >= other.start && self.end <= other.end
     }
 
-    /// returns `true` if this span overlaps with the given one (symetric)
+    /// returns `true` iif this span overlaps with the given one (symetric)
     pub fn overlaps(&self, other: &Span) -> bool {
         !(other.start > self.end || other.end < self.start)
     }
 
-    /// returns `true` if this span has the same offsets as the given one
+    /// returns `true` iif the spans do not overlap
+    pub fn is_disjoint(&self, other: &Span) -> bool {
+        !self.overlaps(other)
+    }
+
+    /// returns `true` iif this span has the same offsets as the given one
     pub fn same_offsets(&self, other: &Span) -> bool {
         self.start == other.start && self.end == other.end
     }